+++ src/Provider/TimeProviderInterface.php
@@ -14,6 +14,8 @@ declare(strict_types=1);
 
 namespace Ramsey\Uuid\Provider;
 
+use Ramsey\Uuid\Type\Time;
+
 /**
  * A time provider retrieves the current time
  */
@@ -22,7 +24,14 @@ interface TimeProviderInterface
     /**
      * Returns a timestamp array
      *
-     * @return int[] Array containing `sec` and `usec` components of a timestamp
+     * @deprecated Transition to {@see TimeProviderInterface::getTime()}.
+     *
+     * @return int[]|string[] Array containing `sec` and `usec` components of a timestamp
      */
     public function currentTime(): array;
+
+    /**
+     * Returns a time object
+     */
+    public function getTime(): Time;
 }