+++ tests/Provider/Time/FixedTimeProviderTest.php
@@ -6,19 +6,23 @@ namespace Ramsey\Uuid\Test\Provider\Time;
 
 use Ramsey\Uuid\Provider\Time\FixedTimeProvider;
 use Ramsey\Uuid\Test\TestCase;
+use Ramsey\Uuid\Type\Time;
 
 class FixedTimeProviderTest extends TestCase
 {
     public function testConstructorRequiresSecAndUsec(): void
     {
         $this->expectException(\InvalidArgumentException::class);
-        $provider = new FixedTimeProvider([]);
+        $this->expectExceptionMessage('Array must contain sec and usec keys.');
+
+        new FixedTimeProvider([]);
     }
 
     public function testCurrentTimeReturnsTimestamp(): void
     {
         $timestamp = ['sec' => 1458844556, 'usec' => 200997];
         $provider = new FixedTimeProvider($timestamp);
+
         $this->assertEquals($timestamp, $provider->currentTime());
     }
 
@@ -32,5 +36,34 @@ class FixedTimeProviderTest extends TestCase
         $provider->setUsec($newTimestamp['usec']);
 
         $this->assertEquals($newTimestamp, $provider->currentTime());
+    }
+
+    public function testGetTimeReturnsTime(): void
+    {
+        $time = new Time(1458844556, 200997);
+        $provider = new FixedTimeProvider($time);
+
+        $this->assertSame($time, $provider->getTime());
+    }
+
+    public function testGetTimeReturnsTimeAfterChange(): void
+    {
+        $time = new Time(1458844556, 200997);
+        $provider = new FixedTimeProvider($time);
+
+        $this->assertSame('1458844556', $provider->getTime()->getSeconds()->toString());
+        $this->assertSame('200997', $provider->getTime()->getMicroSeconds()->toString());
+
+        $provider->setSec(1050804050);
+
+        $this->assertSame('1050804050', $provider->getTime()->getSeconds()->toString());
+        $this->assertSame('200997', $provider->getTime()->getMicroSeconds()->toString());
+
+        $provider->setUsec(30192);
+
+        $this->assertSame('1050804050', $provider->getTime()->getSeconds()->toString());
+        $this->assertSame('30192', $provider->getTime()->getMicroSeconds()->toString());
+
+        $this->assertNotSame($time, $provider->getTime());
     }
 }